requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
decorator==4.4.2
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==25.2.10
fsspec==2024.12.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.7.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
onnxruntime==1.21.0
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyreadline3==3.5.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
setuptools==75.8.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
